services/limbless-app/limbless-server/requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Bcrypt==1.0.1
flask-htmx==0.3.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
libsass==0.22.0
loguru==0.7.2
MarkupSafe==2.1.3
numpy==1.26.0
packaging==23.1
pandas==2.1.0
pluggy==1.3.0
psycopg2==2.9.7
pydantic==1.10.12
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==2.3.7
WTForms==3.0.1
gunicorn==21.2.0